move getLegacyAppealsLength and LEGACY_TYPE into shared and update imports
