+++ src/applications/appeals/995/constants.js
@@ -127,7 +127,6 @@ export const SUPPORTED_BENEFIT_TYPES_LIST = [
   // 'nationalCemeteryAdministration',
 ];
 
-export const LEGACY_TYPE = 'legacyAppeal';
 export const AMA_DATE = '2019-02-19'; // Appeals Modernization Act in effect
 
 export const SUPPORTED_BENEFIT_TYPES = constants.benefitTypes.map(type => ({
